+++ pkg/cmd/machineartifact_test.go
@@ -13,8 +13,8 @@ func TestMachinesArtifactsRetrieve(t *testing.T) {
 		mocktest.TestRunMockTestWithFlags(
 			t,
 			"--api-key", "string",
-			"machines:artifacts", "retrieve",
-			"--machine-id", "machine_id",
+			"machines", "artifacts", "retrieve",
+			"--machine-id", "dm-3",
 			"--artifact-id", "artifact_id",
 		)
 	})
@@ -25,9 +25,9 @@ func TestMachinesArtifactsList(t *testing.T) {
 		mocktest.TestRunMockTestWithFlags(
 			t,
 			"--api-key", "string",
-			"machines:artifacts", "list",
+			"machines", "artifacts", "list",
 			"--max-items", "10",
-			"--machine-id", "machine_id",
+			"--machine-id", "dm-3",
 			"--cursor", "cursor",
 			"--limit", "0",
 		)
@@ -39,8 +39,8 @@ func TestMachinesArtifactsDelete(t *testing.T) {
 		mocktest.TestRunMockTestWithFlags(
 			t,
 			"--api-key", "string",
-			"machines:artifacts", "delete",
-			"--machine-id", "machine_id",
+			"machines", "artifacts", "delete",
+			"--machine-id", "dm-3",
 			"--artifact-id", "artifact_id",
 		)
 	})