+++ pkg/cmd/machineexecution_test.go
@@ -13,8 +13,8 @@ func TestMachinesExecutionsCreate(t *testing.T) {
 		mocktest.TestRunMockTestWithFlags(
 			t,
 			"--api-key", "string",
-			"machines:executions", "create",
-			"--machine-id", "machine_id",
+			"machines", "executions", "create",
+			"--machine-id", "dm-3",
 			"--command", "[string]",
 			"--cwd", "cwd",
 			"--env", "{foo: string}",
@@ -36,8 +36,8 @@ func TestMachinesExecutionsCreate(t *testing.T) {
 		mocktest.TestRunMockTestWithPipeAndFlags(
 			t, pipeData,
 			"--api-key", "string",
-			"machines:executions", "create",
-			"--machine-id", "machine_id",
+			"machines", "executions", "create",
+			"--machine-id", "dm-3",
 		)
 	})
 }
@@ -47,8 +47,8 @@ func TestMachinesExecutionsRetrieve(t *testing.T) {
 		mocktest.TestRunMockTestWithFlags(
 			t,
 			"--api-key", "string",
-			"machines:executions", "retrieve",
-			"--machine-id", "machine_id",
+			"machines", "executions", "retrieve",
+			"--machine-id", "dm-3",
 			"--execution-id", "execution_id",
 		)
 	})
@@ -59,9 +59,9 @@ func TestMachinesExecutionsList(t *testing.T) {
 		mocktest.TestRunMockTestWithFlags(
 			t,
 			"--api-key", "string",
-			"machines:executions", "list",
+			"machines", "executions", "list",
 			"--max-items", "10",
-			"--machine-id", "machine_id",
+			"--machine-id", "dm-3",
 			"--cursor", "cursor",
 			"--limit", "0",
 		)
@@ -73,8 +73,8 @@ func TestMachinesExecutionsDelete(t *testing.T) {
 		mocktest.TestRunMockTestWithFlags(
 			t,
 			"--api-key", "string",
-			"machines:executions", "delete",
-			"--machine-id", "machine_id",
+			"machines", "executions", "delete",
+			"--machine-id", "dm-3",
 			"--execution-id", "execution_id",
 		)
 	})
@@ -85,9 +85,9 @@ func TestMachinesExecutionsEvents(t *testing.T) {
 		mocktest.TestRunMockTestWithFlags(
 			t,
 			"--api-key", "string",
-			"machines:executions", "events",
+			"machines", "executions", "events",
 			"--max-items", "10",
-			"--machine-id", "machine_id",
+			"--machine-id", "dm-3",
 			"--execution-id", "execution_id",
 			"--cursor", "cursor",
 			"--limit", "0",
@@ -100,8 +100,8 @@ func TestMachinesExecutionsOutput(t *testing.T) {
 		mocktest.TestRunMockTestWithFlags(
 			t,
 			"--api-key", "string",
-			"machines:executions", "output",
-			"--machine-id", "machine_id",
+			"machines", "executions", "output",
+			"--machine-id", "dm-3",
 			"--execution-id", "execution_id",
 		)
 	})