+++ pkg/cmd/machinessh_test.go
@@ -13,8 +13,8 @@ func TestMachinesSSHCreate(t *testing.T) {
 		mocktest.TestRunMockTestWithFlags(
 			t,
 			"--api-key", "string",
-			"machines:ssh", "create",
-			"--machine-id", "machine_id",
+			"machines", "ssh", "create",
+			"--machine-id", "dm-3",
 			"--public-key", "public_key",
 		)
 	})
@@ -25,8 +25,8 @@ func TestMachinesSSHCreate(t *testing.T) {
 		mocktest.TestRunMockTestWithPipeAndFlags(
 			t, pipeData,
 			"--api-key", "string",
-			"machines:ssh", "create",
-			"--machine-id", "machine_id",
+			"machines", "ssh", "create",
+			"--machine-id", "dm-3",
 		)
 	})
 }
@@ -36,8 +36,8 @@ func TestMachinesSSHRetrieve(t *testing.T) {
 		mocktest.TestRunMockTestWithFlags(
 			t,
 			"--api-key", "string",
-			"machines:ssh", "retrieve",
-			"--machine-id", "machine_id",
+			"machines", "ssh", "retrieve",
+			"--machine-id", "dm-3",
 			"--session-id", "session_id",
 		)
 	})
@@ -48,9 +48,9 @@ func TestMachinesSSHList(t *testing.T) {
 		mocktest.TestRunMockTestWithFlags(
 			t,
 			"--api-key", "string",
-			"machines:ssh", "list",
+			"machines", "ssh", "list",
 			"--max-items", "10",
-			"--machine-id", "machine_id",
+			"--machine-id", "dm-3",
 			"--cursor", "cursor",
 			"--limit", "0",
 		)
@@ -62,8 +62,8 @@ func TestMachinesSSHDelete(t *testing.T) {
 		mocktest.TestRunMockTestWithFlags(
 			t,
 			"--api-key", "string",
-			"machines:ssh", "delete",
-			"--machine-id", "machine_id",
+			"machines", "ssh", "delete",
+			"--machine-id", "dm-3",
 			"--session-id", "session_id",
 		)
 	})