+++ pkg/cmd/machineterminal_test.go
@@ -13,8 +13,8 @@ func TestMachinesTerminalsCreate(t *testing.T) {
 		mocktest.TestRunMockTestWithFlags(
 			t,
 			"--api-key", "string",
-			"machines:terminals", "create",
-			"--machine-id", "machine_id",
+			"machines", "terminals", "create",
+			"--machine-id", "dm-3",
 			"--height", "0",
 			"--width", "0",
 			"--cwd", "cwd",
@@ -35,8 +35,8 @@ func TestMachinesTerminalsCreate(t *testing.T) {
 		mocktest.TestRunMockTestWithPipeAndFlags(
 			t, pipeData,
 			"--api-key", "string",
-			"machines:terminals", "create",
-			"--machine-id", "machine_id",
+			"machines", "terminals", "create",
+			"--machine-id", "dm-3",
 		)
 	})
 }
@@ -46,8 +46,8 @@ func TestMachinesTerminalsRetrieve(t *testing.T) {
 		mocktest.TestRunMockTestWithFlags(
 			t,
 			"--api-key", "string",
-			"machines:terminals", "retrieve",
-			"--machine-id", "machine_id",
+			"machines", "terminals", "retrieve",
+			"--machine-id", "dm-3",
 			"--terminal-id", "terminal_id",
 		)
 	})
@@ -58,9 +58,9 @@ func TestMachinesTerminalsList(t *testing.T) {
 		mocktest.TestRunMockTestWithFlags(
 			t,
 			"--api-key", "string",
-			"machines:terminals", "list",
+			"machines", "terminals", "list",
 			"--max-items", "10",
-			"--machine-id", "machine_id",
+			"--machine-id", "dm-3",
 			"--cursor", "cursor",
 			"--limit", "0",
 		)
@@ -72,8 +72,8 @@ func TestMachinesTerminalsDelete(t *testing.T) {
 		mocktest.TestRunMockTestWithFlags(
 			t,
 			"--api-key", "string",
-			"machines:terminals", "delete",
-			"--machine-id", "machine_id",
+			"machines", "terminals", "delete",
+			"--machine-id", "dm-3",
 			"--terminal-id", "terminal_id",
 		)
 	})